hack-in support for Cassandra 2.2
Somewhere between 2.1.13, and 2.2.5, Cassandra was refactored to wrap all of
the codahale/dropwizard metric mbeans in Cassandra delegators.  This would
have been awesome if it were done this way from the beginning, but
after-the-fact it broke cmcd by changing the object names of everything.

In order to retain compatibility with the 2.1 series, this changeset
introduces a second collector class (a 2.2-specific one), and uses a factory
(*gasp*) to create instances of either based on the version (which for now
has to be passed along via a system property).

This is unpleasant, but can be backed out rather easily after we no longer
have a need for Cassandra 2.1 compatibility.

Bug: https://phabricator.wikimedia.org/T126629